+++ tomcat/src/main/java/nextstep/jwp/controller/RequestController.java
@@ -1,0 +1,5 @@
+package nextstep.jwp.controller;
+
+public class RequestController {
+
+}
